Transform an expression.

Traverses the expression expr depth-first and transform it using callbacks.

depth_first_rewrite_expr(expr, callbacks, params = list(), topdown = list(),
  wflags = warning_flags(), ...)

Arguments

expr
An R expression

callbacks
List of callbacks to apply.

params
Parameters of the function we are rewriting. If we are working on a raw expression, just use the default, which is an empty list.

topdown
A list of additional information gathered in the traversal.

wflags
Warning flags, see warning_flags().

...
Additional data that will be passed along to callbacks.

Value

A modified expression.

See also